+++ test/integration/src/pages/http/CookieSet/CookieSet.js
@@ -18,7 +18,7 @@ class Component extends React.Component {
       <ul>
         <li>The response should set a 10-years cookie: <code>decade</code> and session cookie <code>one-night-stand</code> if it doesn't exist in request header.</li>
         <li>On second request (when the cookie is already present at request header), response will not set cookie.</li>
-        <li>You can <a href={this.props.router.reverseRoute('COOKIE_REMOVE')}>visit this page</a> to clear cookies from server-side.</li>
+        <li>You can <a href={this.props.context.router.reverseRoute('COOKIE_REMOVE')}>visit this page</a> to clear cookies from server-side.</li>
         <li><a href="javascript:void(0)" onClick={this.setCookies.bind(this)}>Clicking this link</a> would make client-side set 10 years cookie <code>dasawarsa</code> and session cookie <code>cinta-satu-malam</code>.</li>
         <li>Client-side should also be able to <a href="javascript:void(0)" onClick={this.clearCookies.bind(this)}>clear the aforementioned cookies</a>.</li>
       </ul>
@@ -37,20 +37,20 @@ class Component extends React.Component {
 
   readClientCookie() {
     this.setState({
-      'dasawarsa': this.props.cookieJar.read('dasawarsa'),
-      'cinta-satu-malam': this.props.cookieJar.read('cinta-satu-malam')
+      'dasawarsa': this.props.context.cookieJar.read('dasawarsa'),
+      'cinta-satu-malam': this.props.context.cookieJar.read('cinta-satu-malam')
     });
   }
 
   setCookies() {
-    this.props.cookieJar.set(Cookie.createExpireInDays('dasawarsa', '10-tahun', 10 * 360));
-    this.props.cookieJar.set(Cookie.createSession('cinta-satu-malam', 'oh-indahnya'));
+    this.props.context.cookieJar.set(Cookie.createExpireInDays('dasawarsa', '10-tahun', 10 * 360));
+    this.props.context.cookieJar.set(Cookie.createSession('cinta-satu-malam', 'oh-indahnya'));
     this.readClientCookie();
   }
 
   clearCookies() {
-    this.props.cookieJar.remove('dasawarsa');
-    this.props.cookieJar.remove('cinta-satu-malam');
+    this.props.context.cookieJar.remove('dasawarsa');
+    this.props.context.cookieJar.remove('cinta-satu-malam');
     this.readClientCookie();
   }
 }
@@ -75,9 +75,7 @@ class CookieSet extends Page {
     var reactRenderer = new ReactRenderer();
     reactRenderer.head = '<title>Cookie</title>';
     reactRenderer.body = React.createElement(Component, {
-      config: this.config,
-      router: this.router,
-      cookieJar: this.cookieJar,
+      context: this.createContext(store),
       readLifetime: this.cookieJar.read('decade'),
       readSession: this.cookieJar.read('one-night-stand')
     });