+++ test/integration/src/pages/redux/Hydration/Hydration.js
@@ -54,7 +54,7 @@ class Component extends React.Component {
       <h3>Profile 2</h3>
       <UserProfile context={this.props.context} username={'willywonka'}></UserProfile>
       <DebugPanel top right bottom>
-        <DevTools store={this.props.context.reduxStore._store} monitor={LogMonitor} />
+        <DevTools store={this.props.context.store._store} monitor={LogMonitor} />
       </DebugPanel>
     </div>
   }
@@ -84,10 +84,7 @@ class Hydration extends Page {
     var reactRenderer = new ReactRenderer();
     reactRenderer.head = '<title>Hydration Test</title>';
     reactRenderer.body = React.createElement(Component, {
-      context: {
-        reduxStore: store,
-        config: this.config
-      }
+      context: this.createContext(store)
     });
     var renderResult = new RenderResult(reactRenderer);
     callback(renderResult);